+++ changelog.md
@@ -1,1 +1,4 @@
 - fixed copy_from particle types not being parsed as such
+- fixed recursive particles not working
+
+((DAY_TIME%480000)/480000)-0.025"